+++ src/Cabazure.Test/Attributes/InlineAutoNSubstituteDataAttribute.cs
@@ -57,7 +57,7 @@ public sealed class InlineAutoNSubstituteDataAttribute : DataAttribute
     {
         ArgumentNullException.ThrowIfNull(testMethod);
 
-        var fixture = new SutFixture();
+        var fixture = AutoNSubstituteDataHelper.CreateFixture(testMethod);
         var parameters = testMethod.GetParameters();
         var values = AutoNSubstituteDataHelper.MergeValues(fixture, parameters, Values);
         IReadOnlyCollection<ITheoryDataRow> result = [new TheoryDataRow(values)];